tone generator moved to separate subpackage
